Updated Master1 file
The Master1 - Copy.zip file is an updated Master1 file. It has the new "Occupation Factor" information needed to project salaries over time. These factors are provided for each of the three competency levels - Lo, Med, Hi. They are found in the last three columns of the new file. The new file is a compressed Excel file. I'm not sure which format we use so it might need to be converted to a different format.